[Refactor] 매칭카드 무한스크롤 위치 변경 #404
